Fixed a few warnings.
Rev: src/cpp.c:1.68
Rev: src/modules/Image/encodings/png.c:1.36
Rev: src/modules/Image/encodings/tga.c:1.19
Rev: src/modules/_Crypto/crypto.c:1.35
Rev: src/modules/_Crypto/des.c:1.17
Rev: src/modules/_Crypto/idea.c:1.15
Rev: src/modules/_Crypto/invert.c:1.10
Rev: src/modules/_Crypto/md2.c:1.9
Rev: src/modules/_Crypto/md5.c:1.14
Rev: src/modules/_Crypto/nt.c:1.4
Rev: src/modules/_Crypto/sha.c:1.17
Rev: src/modules/spider/spider.c:1.95
Rev: src/modules/spider/stardate.c:1.10
Rev: src/modules/spider/xml.c:1.24
Rev: src/opcodes.c:1.80
Rev: src/preprocessor.h:1.27

+++ src/cpp.c
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: cpp.c,v 1.67 2000/08/08 19:44:06 grubba Exp $
+ * $Id: cpp.c,v 1.68 2000/08/09 13:25:10 grubba Exp $
  */
 #include "global.h"
 #include "stralloc.h"
@@ -480,7 +480,7 @@ while(1)					\
   SKIPSPACE();					\
   if(data[pos]=='/')				\
   {						\
-    INT32 tmp;					\
+    ptrdiff_t tmp;				\
     switch(data[pos+1])				\
     {						\
     case '/':					\
